공통 프로토콜 분리
